backend/requirements.txt
fastapi[all]>=0.116.1
langchain>=0.3.26
langchain-openai>=0.3.28
psycopg2-binary>=2.9.10
python-dotenv>=1.1.1
sqlalchemy>=2.0.41
uvicorn>=0.35.0
requests>=2.31.0